electronapp/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "firebase-admin>=7.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
